sqlite_to_postgres/requirements.txt
attrs==22.2.0
exceptiongroup==1.1.0
flake8==6.0.0
flake8-bugbear==23.1.20
flake8-import-order==0.18.2
iniconfig==2.0.0
isort==5.12.0
mccabe==0.7.0
packaging==23.0
pep8-naming==0.13.3
pluggy==1.0.0
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pyflakes==3.0.1
pytest==7.2.1
python-dotenv==0.21.1
tomli==2.0.1
